refactor: unify test packages, swap acceptance test framework (Phase 5.1a)
Collapses launchdarkly/tests/ back into the root launchdarkly package
and migrates every acceptance / data source test off
terraform-plugin-sdk/v2/helper/{resource,acctest} and /v2/terraform
onto the standalone terraform-plugin-testing fork. Both libraries
register a `sweep` flag in init(), which is why the tests sub-pkg
existed; with the SDKv2 imports gone from the test surface, the root
pkg can host launchdarkly_team_role_mapping alongside everything else.

Same TestCheck / TestStep / TestCase API on both sides — pure import
swap. testAccFrameworkMuxProviders deleted; team_role_mapping now
uses the canonical testAccProtoV5ProviderFactories from
provider_test.go.

Helper-test triage (policy_statements_helper_test.go,
resource_data_helpers_test.go, schema_compat_test.go) deferred to
5.2 — they test SDKv2-typed helpers that haven't been refactored yet.

Refs MIGRATION_PLAN_NON_BREAKING.md Phase 5.1a.
