+++ launchdarkly/resource_team_role_mapping_test.go
@@ -1,7 +1,6 @@
-package tests
+package launchdarkly
 
 import (
-	"context"
 	"fmt"
 	"testing"
 
@@ -106,7 +105,7 @@ func TestAccTeamRoleMapping_basic(t *testing.T) {
 	teamKey := acctest.RandStringFromCharSet(10, acctest.CharSetAlphaNum)
 
 	resource.Test(t, resource.TestCase{
-		ProtoV5ProviderFactories: testAccFrameworkMuxProviders(context.Background(), t),
+		ProtoV5ProviderFactories: testAccProtoV5ProviderFactories,
 		PreCheck:                 func() { testAccPreCheck(t) },
 		Steps: []resource.TestStep{
 			{
@@ -157,7 +156,7 @@ func TestAccTeamRoleMapping_empty(t *testing.T) {
 	teamKey := acctest.RandStringFromCharSet(10, acctest.CharSetAlphaNum)
 
 	resource.Test(t, resource.TestCase{
-		ProtoV5ProviderFactories: testAccFrameworkMuxProviders(context.Background(), t),
+		ProtoV5ProviderFactories: testAccProtoV5ProviderFactories,
 		PreCheck:                 func() { testAccPreCheck(t) },
 		Steps: []resource.TestStep{
 			{
